pyproject.toml
[project]
name = "multimodal-rag-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "ivan",email = "ivanfan@connect.hku.hk"}
]
readme = "README.md"
requires-python = "==3.10.16"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "datasets (>=3.3.2,<4.0.0)",
    "hf-transfer (>=0.1.9,<0.2.0)",
    "ipywidgets (>=8.1.5,<9.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "kagglehub (>=0.3.9,<0.4.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "pypdf (>=5.3.0,<6.0.0)",
    "html2text (>=2024.2.26,<2025.0.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "marker-pdf (>=1.5.5,<2.0.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "einops (>=0.8.1,<0.9.0)",
    "timm (>=1.0.15,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}
